Remove database connections from helper functions
The original code was written with database connections in every function.
The hope was that rewriting and being more conservative with the connections
(i.e. only in the highest level functions), that the memory errors would be
alleviated. This was not the case, but the code is at least cleaner.