Create full location table geocoding workflow (#189)
These changes introduce functions and documentation for a preferred geocoding workflow in which gaia users will geocode their entire OMOP location table and store it as a gaia table. This workflow unlocks the full downstream potential of pairing gaia and OHDSI tools.
Addresses and fixes #185 (the high level issue) and components:
Fixes #188
Fixes #187
Fixes #186
Fixes #180

It does not address some of the outstanding issues associated with Cohort geocoding.